Put `CIName` in it's own file and remove dead constructor
